LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/intversion-gated-on-delete-expected.txt

CONSOLE MESSAGE: line 13: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
Test that a database is recreated correctly when an open-with-version call is queued behind a deleteDatabase call

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-gated-on-delete.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
